#135: move custom colors
