+++ lib/ui/common/noted_library.dart
@@ -6,4 +6,5 @@ export './input/input.dart';
 export './layout/layout.dart';
 export './rich_text/rich_text.dart';
 export './tiles/tiles.dart';
+export './custom_colors.dart';
 export './noted_widget_config.dart';
