+++ noted_catalog/lib/pages/input/catalog_misc_input_page.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:noted_app/theme/custom_colors.dart';
 import 'package:noted_app/ui/common/noted_library.dart';
 import 'package:noted_catalog/catalog_list_widget.dart';
